build,config,doc: Make '../data' default data path, executables in bin
Currently the default data path is './data' so that the binary downloads
just work out of the box. This needs to be changed to '../data' using
the WOLFMUD_DIR environment variable for source downsloads and Git
clones.

This change makes two changes. The first change iconsistently puts all
binaries into the bin directory for all environments. The second change
makes the default data path '../data' for all environments. This means
that setting WOLFMUD_DIR should only be required when setting a custom
path or configuration file.

Details in compiling-from-source.txt and running-the-server.txt in the
docs directory have been updated to reflect the changes.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>