README, docs: Rewritten documentation
This commit is an update of the documentation in the docs directory.
It's main purpose was to document changes required due to the switch to
Go's module mode. However it ended up being a thorough rewrite. Of note
the getting-started.txt file has been removed and sections now included
in the running-the-server.txt and compiling-from-source.txt files.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>