cmd,stats: Collect stats on lock wait times
One of the main metrics that effects how responsive the game is to
players is how long it takes to acquire the locks required.

This commit records the longest lock wait time since the last stat run
and the longest lock wait time since the server started. Note however
that this does not take into account relocking time.

While this change works it has a serious performance penalty due to the
additional contention on the stats channel being used.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>