cmd,stats: Switch from channel to RWMutex for MaxLockWait
The initial implementation of the lock wait statistic has a huge
performance bottleneck due to players waiting for the channel to become
available.

This commit switches to using a RWMutex so that multiple players can
take a read lock when just checking if their wait time is greater than
that recorded. Only if a new wait time needs recording is a read/write
lock then take.

In order for the locking on the RWMutex to be done correctly a new
LockWait function has been added and the relevant variables are now
package local - the channel was previously exported.

This commit also updates the package documentation for the new
statistic and documents some previously omitted values.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>