proc: Replace command switch with a lookp table+rename Inv→Inventory
This commit replaces the command lookup switch in stage.go with a method
lookup map in command.go. It also renames the Inv method to Inventory to
match the long name of the other commands.

Performance is mostly the same except for the case of just gitting enter
and doing nothing which is about 3x slower.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>