proc: Settle on 32bit uint for Is and As values
This commit removes the isAttr and asAttr type aliases and settles on
32bit unsigned integers for both. This allows for 32 Is bit flags and up
to 4,294,967,295 As entries :) Moving from byte to uint32 does not
increase the size of the Thing struct after padding on 32 or 64 bit
platforms, so why not?

Depending on how things develop I might make Is 64bit in the future.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>