proc: Add Container flag to Thing.Is update EXAMINE/PUT/TAKE commands
While every Thing has an In slice for an inventory (which may be nil) it
is not desirable for all Things to be treated as containers. For
example, you should not be able to treat any random item as a container
and PUT/TAKE items from it.

This commit adds a Container flag for Thing.Is so that something can be
specifically identified as a container. If an item is not specifically a
container the PUT/TAKE commands cannot be used on the item. In addition
the EXAM/EXAMINE commands will not show the item's inventory.

The item's inventory may still be manipulated in other ways. For example
a player's In is their main inventory and is not flagged as a container.
The player can still use GET/DROP to add and remove items from their
inventory. However, other player's would not be able to PUT/TAKE another
player's items... maybe STEAL or PICKPOCKET though ;)

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>