proc: Add Thing.As Blocker and Thing.Is Open
This commit adds a new Thing.As constant for a blocker. It's value
should be the direction that is being blocked. This allows a Thing (such
as a door) at a location to block movement in a particular direction.

This commit adds a Thing.Is Open flag to determine if a Thing (such as a
door) is open or closed.

Together Thing.As[Blocker] and This.Is&Open provide functionality for
implementing doors and other barriers to movement.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>