proc: Add Open/Close for doors, update Move/Examine
This commit adds the OPEN and CLOSE commands for doors and other
barriers to movement. For a door to be functional it must have a
Thing.As[Blocker] set to the direction being blocked and a
This.As[Where] so it knows its primary location. For example:

  door := proc.NewThing(
        "the tavern door",
        "This is a sturdy wooden door with a simple latch.",
  )
  door.Is = proc.Narrative
  door.As[proc.Alias] = "DOOR"
  door.As[proc.Blocker] = "E"
  door.As[proc.Where] = "L3"

Currently a door/barrier needs to be added to the to and from locations
on either side of the door/barrier. This should be done automatically in
a future update.

The state.Move method has been updated to implement the main
functionality blocking movement. The state.Examine method has been
updated to indicate whether the door/barrier is currently open or
closed.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>