proc,cmd: pass in io.Writer for state to write back on
Currently proc.stat writes output back to the player using fmt.Printf
and fmt.Println to write to the default Stdout.

This commit updates NewState to take an io.Write parameter and to use
that to write back to the client. In future this can then be a buffer or
a network connection.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>