proc: Switch back to Quit function and clean up quit processing
This commit switches back to using the Quit method for the QUIT player
command. It is going to be needed later on for removing the player from
the world and saving the player data.

This commit updates the Quit method to provide a better message and
clears the current prompt so nothing is displayed after the quit
message.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>