proc,world: Comment proc.World+world.Load mention crosstalk as BUG
This commit documents proc.World and world.Load properly. It also
comments a BUG as world.Load dumps the world directly into proc.World
which is poor form - proc.World is populated as a side effect of
world.Load which isn't even in the same package! I can get away with
this for now, but it really needs to be sorted out properly...

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>