proc,cmd,world: Move Thing Name+Description into Thing.As
Originally I thought as everything is likely to have a name and
description, and that name will be used a lot, it would be better to
have them as Thing.Name and Thing.Description instead of having them as
Thing.As[Name] and Thing.As[Description].

However, while working on the world zone loader it became apparent that:

  a) Using NewThing(name, description string) was a pain
  b) Having name and description as a special case was a pain
  c) Having everything as properties was much nicer to work with.

This commit therefore updates Thing to remove the Name and Description
fields and adds As constants for Name and Description. This commit also
updates the Thing.dump method to handle wrapping any long 'As' lines.
Finally, this commit updates the static world builder and simple player
creation to use Thing.As[Name] and Thing.As[Description].

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>