proc: Stop being lazy with Is/As names and constants
This commit cleans up the Is/As constants and isName/asName definitions.
Instead of being lazy and compressing everything to a few lines put each
item on it's own line. This allows the lists to be organised and
add/removing an entry should not cause a larger than necessary diff.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>