proc: Add Thing.Free method
The Thing.Free method is not necessary, however testing has shown that
it can aid the garbage collector. When using Free objects are finalised
much more quickly and in larger groups. Without Free the garbage
collector seems to take several passes in order to work which objects
are no longer alive and can be disposed of.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>