proc: Use named types for Is/As indexes, make UID uint
At the moment it is too easy to accidentally mix up Is and As constants
resulting in awkward runtime bugs. This commit switches to named types
for the Is and As indexes resulting in compile time errors if Is/As
indexes are used incorrectly.

This commit also updates the UID counter to be uint as there is no
reason for it to be uint32.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>