proc,cmd,world: Add Thing.Any, Move proc.Alias to Thing.Any
Currently attributes of a Thing are bit flags (Thing.Is) or single
strings (Thing.As). However, for attributes such as aliases and vetoes
we need multiple keys/values. Aliases need multiple values for the
alias, qualifiers and variations. Vetoes need multiple keys, one for
each command being vetoed.

Vetoes could be implemented using Thing.Is by defining multiple keys:
VETO_GET, VETO_PUT etc. However, this could lead to a large numder of
constants. A more programmer friendly solution seems to be dynamic keys
such as defining:

  const (
    Alias anyKey "ALIAS"
    Veto  anyKey "VETO:"
  )

Here Veto is a compposite key because it has a ':' suffix. It can be
used as:

  item.Any[VETO+"GET"]

This is the approache taken by this commit, although vetoes are not
defined or implemented yet.

This commit implements Thing.Any:

  - Adds Thing.Any as map[string][]string and defines Alias as an
    anyKey, dropping the previous Alias asKey definition.

  - The dump method has been updated to include Thing.Any information.

  - NewThing has been updated to initialise the Thing.Any map.

  - Thing.Copy has been updated to include Thing.Any in the copy.

  - The Find function has been updated to look at all aliases for a
    match. BUG: This is currently incorrect as it matches qualifiers and
    bound qualifiers as simple aliases for now.

  - main.go and world.go have been updated to place aliases into
    Thing.As when setting up a Thing.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>