proc,world: Move vetoes from Thing.Any to Thing.As + add more vetoes
After experimenting with vetoes some more they appear to fit more
naturally into being a Thing.As than a Thing.Any value. Although I'm
still not too sure about defining all of the Veto* constants which seems
to be a little messy...

This commit changes vetoes to be a Thing.As instead of Thing.Any value.
The world definitions have been updated and vetoes for the fireplace and
fire added. Currently implemented vetoes are:

        Get
        Drop
        Put
        PutIn (pseudo command for containers)
        Take
        Takeout (pseudo command for containers)

Vetoes are currently only active for items and containers. Putting a
veto on a location or NPC will not cause the location or NPC to veto a
command occurring in its surroundings. This means, for example, peaceful
locations could not be created yet - and there is no combat yet...

Actually thinking about the Veto* constants, this is going to be a pain
to handle in the loader as we'll need a switch to match the recordjar
data to the correct veto. For example:

  Veto: JUNK→The lattice cannot be junked.

  switch veto {
  case "JUNK":
    item.As[VetoJunk] = vetoMsg
  :
  :
  :
  default:
    log.Printf("Unknown veto: %s, for: %s", veto, item.As[Name])
  }

Maybe another map is needed to map veto strings to veto constants? That
would work...

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>