proc: Add Location to Thing.Is bit flags
Currently the only way to identify a Thing as a location is to check
Thing.As and see if any of the direction constants have been set. This
change adds Location to the Thing.Is bit flags. Setting the flag
indicates the Thing should be treated as a location. This makes it
quicker to check for a location and allows for locations without
predefined exits.

At the moment this flag is not set, but will be set by the new loader.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>