world,proc,cmd: New zone file based world loader (WIP)
This commit replaces the hard-coded, temporary world loader with one
that can create the world by loading existing zone files from the
current WolfMUD. While in a working state, the zone loader is still a
work in progress and incomplete.

This commit implements the multiple starting locations and picks one at
random for the player.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>