proc,world: Change Thing.In from []*Thing to map[String]*Thing
Currently the inventory for a Thing is stored as a []*Thing. This means
to find a specific item we always have to range over the inventory. We
need to find specific items a lot. It also means that removing inventory
items is more complex than it needs to be. The slice elements have to be
copied backwards one element and the extra element at the end set to
nil.

It does have the nice property of keeping inventories in a predictable
order with items always added to the end of the list. However, this
benefit is not outweighed by the issues.

This commit changes the Thing.In to be a map[String]*Thing where the key
is the UID of the item in the inventory. It makes inventory management
simpler and accessing a specific thing by UID much quicker. To help with
the ordering of inventories (mainly the LOOK, INVENTORY and EXAMINE
commands) a new Thing.SortedIn method has been added that returns the
content of the inventory as a slice of *Thing ordered by UID.

Currently NewThing always create the map for all Thing.In which could
waste storage. It would be possible to set Thing.In to nil for a Thing
that is neither a location or a container. However, I don't really want
to be chasing stray nil pointer panics at the moment :(

Command effected by this change have been updated - much code deleted :)
The world loaded has also been updated to populate the inventories using
the item UID.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>