proc: Add item matcher
This commit adds new proc.Match and proc.LimitedMatch methods that
matches a list of words against a list of items - matching on the item's
aliases, qualifiers and bound qualifiers. A list of matched UIDs and/or
unmatched words is returned.

All existing player commands need updating to use proc.Match and/or
proc.LimitedMatch where applicable.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>