proc: Clean up handling messages to actor (Msg, new MsgAppend)
Currently messages are built add-hock with line-feeds being littered
everywhere.

This commit updates and comments the state.Msg method to automatically
add line-feeds before a message if nothing has been written to the actor
yet. This commit also adds a new state.MsgAppend which never adds
line-feeds automatically. This can be useful when building a message in
stages.

Finally this commit cleans up the message in the currently implemented
commands. A bug is also noted for proc.Examine where inventory items are
always listed - even if it is only a single item.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>