proc,cmd: Actually put the player in the game
Currently the player just holds a reference to the location they are in
and we work out everything from there. This means that a player is never
actually in the world, the are never placed in a location's inventory.
This will cause a problem later on when other players want to see who is
at the same location as them.

This commit updates the initialisation of the player in main to put the
player in the world. It also updates the MOVE and TELEPORT commands to
remove the player from the source location's inventory and add it to the
destination location's inventory. The LOOK command has been updated to
ommit the current actor from the list of items shown at a location.

Finally, the QUIT command removed the player from the world.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>