proc,core: Rename proc package as core
As a name 'proc' seemed okay for the main package. It was short for
'process' and was supposed to indicate that it handled all of the main
processing for the game. Having worked with the code a lot more, 'proc'
now seems inadequate for the package name.

This commit changes the package name to 'core' to indicate that it is
the core package for the game. The core package has the core types and
core functionality. The name is also short without having to be an
abbreviation.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>