mailbox: new mailbox package for messaging players
The mailbox package provides asynchronous messaging for players. A
player calls Add to add a new mailbox and gets a receive only channel on
which messages are delivered.

Messages are sent to a player by calling Send with the UID of the
recipient. Messages are queued in a channel until received by the
recipient. If the queue is full the oldest message will be dropped to
make room for the new message.

In this way messages are queued for players with slower connections
until they can catch up. It also means that player commands can fire off
messages without being delayed by the network waiting for messages to be
sent and received.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>