cmd: Fix botrunner sending an empty action after every action
In the send method a "\r\n" is always added to the data being sent.
However, in the do method we are adding "\r\n" to the action before
sending the action via the send method. This results in an empty action
being sent after every action:

  SAY Hi!\r\n\r\n

This commit removes the extraneous "\r\n" that is being concatenated to
the action in the do method.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>