cmd: Switch from net.Conn to net.TCPConn
Currently the server uses net.Listen and net.Accept which produces a
net.Conn for the player's network connection. However, this does not
provide fine grained control of the connection.

This commit switches to using net.ListenTCP, net.AcceptTCP which gives
us a net.TCPConn for the player connection, which has more knobs we can
twiddle.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>