core,cmd: Switch player to using mailboxes
Currently an io.Writer is passed into state.NewState and state.Parse
writes to the io.Writer. In effect state is writing directly to the
player's network connection. With this model we have no way to message
other players, for example to notify them of something happening. If we
could and the other player had a slower connection it would slow down
processing for the actor player and other players being notified.

This commit updates state.NewState to no longer take an io.Writer.
Instead state.Parse uses the new mailbox facility to send a message to
the player asynchronously.

This commit updates game.player to poll the player's mailbox for new
messages. It is then game.player that writes messages to the player's
network connection. This then means that all of the network code is now
in cmd/game/main.go, but does mean we have two goroutines per player
instead of one.

For now, prompt handling is all in game.player which means the QUIT
player command can no longer remove the prompt and we get an extra
prompt displayed when the player quits :(

With these changes in place it is now possible to implement player
commands like SAY that send messages to other players. It is also
possible to update existing commands like LOOK, GET, DROP, etc to inform
other players of a player's actions :)

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>