core: Add a Thing.Who field to hold players @ location
Currently everything at a location including narratives, items, NPCs and
players are all held in the Thing.In field. However, while the players
at a location change a lot as the players move around the rest of the
inventory changes a lot less.

This commit adds a Thing.Who field that will contain the players at a
location. This makes it quick and easy to identify players/ono-players
instead of having to filter everything at a location. It will also allow
us to manage the map used to hold the players more efficiently in the
future if needed.

This commit also update the thing.dump method to include thing.Who
information.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>