core: Refactor Thing.SortedIn→Things.Sort, new Things type
Currently the thing.SortedIn method only works on the Thing.In map. Now
we habe Thing.Who a more general way of sorting map[string]*Thing is
needed.

This commit adds a new type Things for map[string]*Thing so that we can
represent the map as a named type. The Thing.SortedIn method has then
been reimplemented as the Things.Sort method. In the future additional
helpers could be added to the Things type.

The Thing.In and Thing.Who fields are now Things. The core.World has
also been changed to be Things.

Calls to SortedIn has been replaced by calls to Sort.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>