core: Add crowd control to matcher
When matching aliases we check Thing.In and Thing.Who. However, if there
is a crowd of players then instead of a list of players we will see:

  It's too crowded to see anything here.

This means we won't be able to single out a player. If we can't single
out a player there is no point checking against all the players.

This commit ommits players from the matcher.match function when matching
in a crowded location.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>