core: use multiple message buffers in core.state+refactor commands
Currently core.state uses a single message buffer which is for the
actor. Messages can be sent directly to participants or observers using
mailbox.Send, however there is no mechanism for batching together
multiple messages for the participants and observers. This could result
in, for example, "GET BAG SACK APPLE" being seen as:

  >
  You see Diddymus get a bag.
  >
  You see Diddymus get a sack.
  >
  You see Diddymus get an apple.
  >

Preferably this should be seen as:

  >
  You see Diddymus get a bag.
  You see Diddymus get a sack.
  You see Diddymus get an apple.
  >

This commit changes the actor's single message buffer to be multiple
message buffers. The state.Msg and state.MsgAppend methods have been
updated to accept a recipient. The recipient can be a player, usually
the actor and maybe a participant, or a location. Messages are first
sent to individual players then to players at the locations. If a player
receives an individual message it will not receive the less specific
location message, Any number of players can be specified as can multiple
locations.

All usage of state.Msg and state.MsgAppend in the player commands have
been updated.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>