cmd,core: Add Thing.As[DynamicAlias]
Currently there is an issue when the matcher is used with the alias
PLAYER in conjunction with the LOOK command. Assume we have to players
in the same location: PLAYER1 and PLAYER2. PLAYER1 uses the LOOK command
and it shows:

  >LOOK
  [Room]
  You are in a small room.

  You see PLAYER2 here.

  You see exits: East
  >EXAMINE PLAYER
  You examine PLAYER1
  This is adventurer, just like you.
  >

The reason for this is that the location Thing.Who has both players
causing the matcher to match on the first player in the list, PLAYER1.
However, the LOOK command is filtering out the current player and not
showing the full list. This can be a little confusing.

This commit adds a new Thing.As[DynamicAlias]. It is used to supplement
the normal aliases in Thing.Any[Alias]. For a player this is normally
set to PLAYER and acts and matches as a normal PLAYER alias would.

However, when command is parsed for a player (or any other actor) the
dynamic alias is changed to SELF for the duration of the command and
then set back to it's previous value.

This has two effects. Firstly, the actor is not included in the matches
for PLAYER for their own commands and so the above example works as
expected. Secondly, the actor can always refer to themselves as SELF:
EXAMINE SELF, #DUMP SELF, TALK SELF Where did I put...

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>