cmd,core: Add Thing.As[DynamicQualifier] "MY"
This is an idea that came from implementing the DynamicAlias.

Currently there is a situation where an item with the same alias exists
in the location and in a players inventory. Then depending on the
command being used the player may have to use an index in order to
interact with the desired item. For example:

  [Bladesmith workshop]
  You are in a small workshop that makes and sells weapons. Against the
  walls are racks holding large swords and polearms. Smaller hand
  weapons are displayed on some trestle tables.

  You see a battle axe here.
  You see the bladesmith here.
  You see a shortsword here.
  You see a runestone here.
  You see a dagger here.
  You see a longsword here.

  You see exits: North
  >get sword
  You get a shortsword.
  >examine sword
  You examine a longsword.
  This is a fine longsword. It has a pointy end.
  >examine 2 sword
  You examine a shortsword.
  This is a fine shortsword. It has a pointy end.
  >

This commit adds Thing.As[DynamicQualifier]. When set is is used to
supplement the normal qualifiers in Thing.As[Qualifier]. When the
GET/TAKE commands are used the item is given a dynamic qualifier of
"MY". When the DROP/PUT commands are used the dynamic qualifier is
removed. This has the effect of allowing a player to refer to any item
in their inventory using "MY". The above example then becomes:

  [Bladesmith workshop]
  You are in a small workshop that makes and sells weapons. Against the walls
  are racks holding large swords and polearms. Smaller hand weapons are
  displayed on some trestle tables.

  You see a battle axe here.
  You see the bladesmith here.
  You see a shortsword here.
  You see a runestone here.
  You see a dagger here.
  You see a longsword here.

  You see exits: North
  >get sword
  You get a shortsword.
  >examine my sword
  You examine a shortsword.
  This is a fine shortsword. It has a pointy end.
  >

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>