core: Drop items and notify observers when QUIT used
Currently if a player quits while carrying items the items are lost as
item resetting and respawning has not been implemented yet. When a
player does quit they simply disappear without notice.

This commit forces a player to drop all items when quitting so that the
items are not 'lost' from the world.

This commit also notifies observers when a player quits and leaves.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>