core: Fix GET/DROP not notifying observers, don't use >= CrowdSize
In the GET and DROP commands crowded is checked as '>= CrowdSize'.

However, we then use '!crowded' to test if the location is crowded, when
we actually should be using 'crowded'. Using '>= CrowdSize' when we
mostly use '< CrowdSize' feels clumsy and easy to get wrong, if we use
'> CrowdSize' it would be off by one.

This commit switches to using '< CrowdSize' throughout. We then invert
the negative 'crowded' test to a positive 'notify' test instead.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>