core: Move Thing types into core/types.go and clean-up
Currently core/thing.go is very cluttered with the type definitions and
lookup tables for the types used by Thing. However, this is swamping the
definition of Thing itself.

This commit moves the ancillary types into a new core/types.go source.
The definitions in the new types.go file have been cleaned-up and
comments improved.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>