core,cmd: command magic init dance move to RegisterCommandHandlers
Currently command handlers are registered during the package variable
initialisation of core/command.go, except for the COMMANDS command which
is added via an init function - otherwise it generates a cyclic
initialisation error. A similar situation has arisen while trying to add
the $ACTION scripting command for action events.

This commit removes all of the 'magic' init dance. Instead a new
function RegisterCommandHandlers has been added to perform all of the
required initialisation. All that is needed is to call the function when
the server starts.

The COMMANDS command has been updated to to renaming of the command list
from cmdNames to commandNames. state.parse has been updated due to
commands being renamed commandHandlers.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>