core: Add state.subparse and move BWL+mailman parse→Parse
This commit adds a new subparse method to state. The method is intended
to allow commands to call other commands by passing an input string. For
example "GET #UID-BAD". The sub-command is processed with the same actor
and buffers as the original command, allowing messages to be accumulated
across multiple commands.

The subparse method calls parse, which will already have locked the BWL.
To account for this locking/unlocking the BWL has been moved to the
Parse method. This allows parse and subparse to be called as many times
as required. The call to mailman to deliver messages has also been moved
to the Parse method. This allows all messages accumulated by parse and
subparse to be sent as a single batched response to each recipient.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>