core: Enable MOVE command to work with non-players
Currently the MOVE command expects actors to be players that are found
in the Thing.Who slice. However this is not the case for e.g. NPCs.

Currently it is undecided as to whether the NPCs should be moved into
Thing.Who or kept in Thing.In - this could later complicate non-NPC, for
example vehicles, movement.

This commit handles the case of moving non-players from one Thing.In to
another Thing.in - the logic is the same as for moving a player from one
Thing.Who to another Thing.Who except we don't need to use the LOOK
command after moving in order to describe the new location.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>