core: Add eventKey type and definitions
This commit adds the eventKey type for use in Thing.Events. It should be
noted that the key's value is actually a pointer into Thing.Int and we
assume that the Thing.Int After value is the same as the eventKey and
that the Thing.Int Jitter value is eventKey+1. This means that the After
and Jitter Thing.Int keys must be consecutive.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>