core, world: Add Thing.Enable and use from World.Load (fix door properly)
Currently World.Load does a lot of setup for items placed in the world.
However, it would be preferable to keep World.Load was as dumb as
possible. Having the setup for items outside of the loader also makes it
reusable for setting up spawned and/or reset items.

This commit adds a Thing.Enable method to contain all of the setup logic
for an item placed in the world. The logic for setting up blockers, such
as doors, has been moved from World.Load to Thing.Enable.

This command also fixes doors so that work properly again without nasty
hacks.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>