core: Disable action events on GET, re-enable on DROP
Currently if an object is picked up its action events keep on going.

This commit aborts action events when something is picked up and
re-enables them when the item is dropped again. We only need to handle
GET and DROP as PUT and TAKE requires the item to move from or to the
actors inventory.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>