core,world: Add support for disabled, out of play, items
This commit adds support for a Thing to be able to hold disabled, out of
play, items. It also adds support for loading items that are initially
disabled and out of play.

This functionality could have been implemented by adding a Thing.Is
Disabled bit flag. However, this would mean any code handling Thing.In
would have to check for disabled items.

Instead this commit adds a new Thing.Out inventory for holding disabled,
out of play, items. This means disabled items are not even seen by code
handling Thing.In items. The trade-off for simpler code is that Thing
incurs storage for an additional map.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>