core: Add Thing.Ref
This commit adds Thing.Ref for holding references to other Thing by key.
For example: Where - where a thing is, Origin - where a Thing should
reset to.

This could also be used to replace location soft linking by UID with a
reference to the location itself. For example:

  Thing.As[<direction>] = string

Becomes:

  Thing.Ref[<direction>] = *Thing

This would let us has more complex location topology. For example,
locations in locations.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>