core: Add radius helper function - calculates locations within N moves
This commit adds a radius function that returns all locations within N
moves of a given location. The locations are returned as a [][]*Thing
where the first index is the number of moves from the central location,
which is always at [0][0]. For example [1][...] contains all locations
within one move of the given location.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>