mailbox,core: Add message priority + message despamming
This commit adds a simple flag to indicate if a message is a priority
message or not. Priority messages are always sent. Non-priority messages
are only sent when the message is not the same as the previously sent
non-priority message. This is useful in preventing repeated streams of
"You hear talking nearby." or "You hear a sneeze.".

The state.mailman has been updated to send all of an actor's messages to
themselves as priority - an actor always get all of their own, even
repeated, messages. All other messages are non-priority.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>