core: Drop World usage from state
Currently state stores message buffers via a map[string]*strings.Builder
using the UID of the recipient or location. With the recent changes to
preferring Thing.Ref links, instead of Thing.As[UID], we could end up
with the recipient being somewhere we cannot message due to a World[uid]
lookup failing.

This commit switches the buffers to be map[*Thing]*strings.Builder and
drops accessing World. This also seems to be more efficient: 110%-112%
usage compared to 113%-115% CPU usage with 64,000 bots on my i5-2400.

The commands DUMP, LOOK, PUT and TAKE which access the buffers directly
(tut-tut!) have been updated.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>