core: Add $RESET command to handle resetting an item
The $RESET command will place a currently out of play item (from
Thing.Out) back in to the world (to Thing.In). If an inventory has WAIT
set an it contains out of play items it will not reset. Instead the last
out of play item to reset will schedule the inventory to reset. This
enables containers to only reset when all of their inventory is ready.

If an OnReset message has been provided it will be used. Otherwise a
default message will be used. If an empty OnReset message is provided
then the default message will be supressed. If a reset occurs in a
container the message will be sent to where the container is.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>