core: Add Thing.Spawn to create spawnable things
This commit adds two methods: a Thing.Spawn wrapper and Thing.spawn.

Thing.Spawn can be used to create a spawned copy of a Thing, complete
with any inventory items, the original item is put out of play and a
reset scheduled for it.

For the spawned copy, inventory items retain their spawnability and any
non-spawnable items are moved from the original to the copy. This lets
us have containers with items where the player can put additional items
in the container before picking up the container. In this case the
additional items would be moved to the player's spawned copy.

If Spawn is called with a non-spawnable Thing the original Thing is
returned.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>