core: Add a simple JUNK player command
This commit adds a simple JUNK command in order to be able to dispose of
all the items we can now spawn ;)

NOTE: We are not handling unique items (such as mobiles) yet. If you
junk, for example, the frog it is gone for good - at least until the
server is restarted.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>