core: type added for Thing.Ref[Origin] used by unique item resets
Unlike spawnable items, theere is only ever one instance of a unique
item (although more than one instance can be defined, but there will
only be one of each instance in the world). For spawnable items it is
always the spawnable copy the player receives. For unique items it is
the single instance that is moved around. When a unique items needs to
be reset it must be moved back to its original location. However, the
original location for unique items is not currently recorded.

This commit adds the refKey Origin constant and name for use as
Thing.Ref[Origin] to record a unique item's origin. The Thing.Enable
method has been updated to record a unique item's origin.

Signed-off-by: Andrew 'Diddymus' Rolfe <diddymus@wolfmud.org>